+++ src/components/admin/TaskForAdmin.vue
@@ -51,7 +51,7 @@
 </template>
 
 <script>
-import axios from "axios";
+import axios from "axios/index";
 
 export default {
   name: "TaskForAdmin",
