+++ src/components/public/Footer.vue
@@ -1,0 +1,17 @@
+<template>
+  <div style="color: white; height: 100%; text-align: center" id="footer">
+    <div style="padding-top: 20px">By XLab 众测小组</div>
+  </div>
+</template>
+
+<script>
+  export default {
+    name: "Footer"
+  }
+</script>
+
+<style scoped>
+  #footer{
+    background: url("https://cdn.luogu.org/fe/footer-bg.png?c32d69c2308cb4267b8bd6fafadddd32");
+  }
+</style>
