+++ src/components/public/WorkerAsideNav.vue
@@ -1,0 +1,15 @@
+<template>
+  <div style="background-color: #34495e; height: 100%">
+    <div style="color: white">Aside,不同身份有不同aside</div>
+  </div>
+</template>
+
+<script>
+  export default {
+    name: "WorkerAsideNav"
+  }
+</script>
+
+<style scoped>
+
+</style>
