+++ src/components/questions/QuestionVer1.vue
@@ -1,22 +1,26 @@
 <template>
   <div>
     <el-row>
-      <el-col :span="12">
+      <el-col :span="12" :offset="2">
         <el-row>
           <el-col :span="19" :offset="5">
-            <div style="font-size:16pt;word-wrap:break-word;padding-top:2pt">{{qtmp.desc}}</div>
+            <div style="font-size:16pt;word-wrap:break-word;padding-top:2pt">
+              <strong>{{qtmp.desc}}</strong>
+            </div>
           </el-col>
         </el-row>
         <el-row style="margin-top: 20px">
-          <div style="height: 300px; overflow: hidden">
-              <img style="width: 400px" :src="qtmp.url">
+          <div style="">
+              <img style="width: 400px; height: 400px" :src="qtmp.url">
           </div>
         </el-row>
       </el-col>
-      <el-col :span="12">
-        <div style="font-size:18pt; text-align:center">选项</div>
+      <el-col :span="8">
+        <div style="font-size:18pt; text-align:center">
+          <strong>选项</strong>
+        </div>
         <el-radio-group v-model="opt" style="width: 100%">
-          <el-row v-for="(item,index) in qtmp.opts" :key="index" style="margin-top:12px">
+          <el-row v-for="(item,index) in qtmp.opts" :key="index" style="margin-top:30px">
             <el-col :span="9" :offset="3" style="text-align:left">
               <el-radio :label="index" style="font-size:20pt">
                 <span style="font-size:16pt">{{item.content}}</span>
