+++ src/components/questions/QuestionVer3.vue
@@ -1,10 +1,12 @@
 <template>
   <div>
     <el-row>
-      <el-col :span="12" :offset="4">
+      <el-col :span="12" :offset="2">
         <el-row>
-          <el-col :span="19" :offset="3">
-            <div style="font-size:16pt;word-wrap:break-word;padding-top:2pt">{{qtmp.desc}}</div>
+          <el-col :span="19" :offset="5">
+            <div style="font-size:16pt;word-wrap:break-word;padding-top:2pt">
+              <strong>{{qtmp.desc}}</strong>
+            </div>
           </el-col>
         </el-row>
         <div style="width: 420px; height: 420px; position: relative; margin-top:20px">
@@ -23,10 +25,12 @@
         </div>
       </el-col>
       <el-col :span="8">
-        <div style="font-size:18pt; text-align:center">类别</div>
+        <div style="font-size:18pt; text-align:center">
+          <strong>类别</strong>
+        </div>
         <el-radio-group v-model="opt" style="width: 100%" :fill="qtmp.opts[opt].color">
           <el-row v-for="(item,index) in qtmp.opts" :key="index" style="margin-top:30px">
-            <el-col :span="9" :offset="9" style="text-align:left">
+            <el-col :span="9" :offset="3" style="text-align:left">
               <el-radio-button
                 :label="index"
                 style="font-size:20pt; padding:0; padding-top:5px; padding-left:3px; padding-right:3px"
@@ -92,7 +96,7 @@ export default {
       let tem={
         index:this.qtmp.index,
         ans:this.pointsCollection
-      }
+      };
       return tem;
     },
     // cancel() {
