+++ src/components/requester/TaskForRequester.vue
@@ -36,7 +36,7 @@
           <el-button type="primary" @click="getAns">查看答案</el-button>
         </router-link>
         <router-link v-bind:to="'/admin-task-detail/' + task.id" target="_blank">
-          <el-button type="primary" @click="dialogVisible = false;">查看详情</el-button>
+          <el-button type="primary" @click="dialogVisible = false">查看详情</el-button>
         </router-link>
       </span>
     </el-dialog>
@@ -44,7 +44,7 @@
 </template>
 
 <script>
-import axios from "axios";
+import axios from "axios/index";
 
 export default {
   name: "TaskForAdmin",
