+++ src/components/requester/projectEditor/QuestionListPreview.vue
@@ -2,41 +2,69 @@
   <div style="width:75%; margin:auto">
     <h1 style="text-align:center">{{taskInfo.name}}</h1>
     <div
-      v-for="(item,index) in tmpList"
+      v-for="(item,index) in showedQuestions"
       :key="index"
       style="margin-top:50px; border:blue 1px solid; padding:20px; border-radius:10px"
     >
       <component :is="qtype" :qtmp="item" :ref="index"></component>
     </div>
+    <p style="text-align:center"><el-pagination
+      background
+      layout="prev, pager, next"
+      :page-size="pageSize"
+      :total="tmpList.length"
+      :current-page.sync="curPage"
+      @current-change="handleCurrentChange"
+    ></el-pagination></p>
     <div style="text-align:right; margin-top:20px">
-      <el-button type="primary" @click="submitAnswer">提交答案</el-button>
-      <el-button @click="test">测试</el-button>
+      <el-button type="primary" @click="submitAnswer">模拟提交</el-button>
     </div>
   </div>
 </template>
 
 <script>
-import QuestionVer1 from "@/components/QuestionVer1";
-import QuestionVer2 from "@/components/QuestionVer2";
-import QuestionVer3 from "@/components/QuestionVer3";
-import QuestionVer4 from "@/components/QuestionVer4";
-import axios from "axios";
+import QuestionVer1 from "@/components/questions/QuestionVer1";
+import QuestionVer2 from "@/components/questions/QuestionVer2";
+import QuestionVer3 from "@/components/questions/QuestionVer3";
+import QuestionVer4 from "@/components/questions/QuestionVer4";
+import QuestionVer5 from "@/components/questions/QuestionVer5";
+import axios from "axios/index";
 export default {
+  props: {
+    tid: Number
+  },
   data() {
     return {
+      //tid:83,
       qtype: "",
       tmpList: [],
       ansList: [],
-      taskInfo: {}
+      taskInfo: {},
+      pageSize:5,
+      curPage:1
     };
   },
   components: {
     QuestionVer1,
     QuestionVer2,
     QuestionVer3,
-    QuestionVer4
+    QuestionVer4,
+    QuestionVer5
+  },
+  computed: {
+    showedQuestions() {
+      let fst=this.pageSize*(this.curPage-1);
+      let lst=this.pageSize*this.curPage;
+      let tem=this.tmpList.slice(fst,lst);
+      
+      return tem;
+    }
   },
   methods: {
+    handleCurrentChange(val) {
+      this.curPage = val;
+      //alert(val);
+    },
     dateToString(draftTimeV) {
       draftTimeV = draftTimeV + "";
       let date = "";
@@ -63,44 +91,29 @@ export default {
         let tem = this.$refs[i][0].getAns();
         this.ansList.push(tem);
       }
-      // console.log(this.ansList);
-      // console.log(this.tmpList);
-      let date = new Date();
-      let dateStr = this.dateToString(date);
-      let para = new URLSearchParams();
-      para.append("taskId", this.taskInfo.id);
-      para.append("answerTime", dateStr);
-      para.append("answer", JSON.stringify(this.ansList));
+      //console.log(this.ansList);
+    },
+    test() {
       axios
-        .post("/api/answer/add", para)
+        .get("/api/answer/find-my-answer")
         .then(response => {
-          alert("提交成功");
           //console.log(response.data);
-          console.log(response.data);
         })
         .catch(response => {
           alert("error");
         });
-    },
-    test(){
-      axios.get('/api/answer/find-my-answer').then(response=>{
-        console.log(response.data);
-      }).catch(response=>{
-        alert('error');
-      })
     }
   },
   mounted() {
-    let tid = this.$route.query.tid;
-    alert(tid);
+    //alert(this.tid);
     axios
-      .get("/api/task/find-by-id", { params: { id: tid } })
+      .get("/api/task/find-by-id", { params: { id: this.tid } })
       .then(response => {
         // console.log(response.data);
         this.qtype = "question" + "-" + response.data.task.type;
         this.taskInfo = response.data.task;
         axios
-          .get("/api/task/read-resource", { params: { taskId: tid } })
+          .get("/api/task/read-resource", { params: { taskId: this.tid } })
           .then(response => {
             let res = response.data;
             let urls = res.urls;
@@ -115,7 +128,16 @@ export default {
                   text: urls[i].text,
                   index: urls[i].index
                 };
-              } else {
+              } else if(this.qtype=="question-ver5"){
+                tmp={
+                  desc: desc,
+                  opts: opts,
+                  context: urls[i].context,
+                  intent: urls[i].intent,
+                  index: urls[i].index
+                };
+              }
+              else {
                 tmp = {
                   desc: desc,
                   opts: opts,
