+++ src/components/worker/ReceivedTaskfForWorker.vue
@@ -30,7 +30,7 @@
 </template>
 
 <script>
-import axios from "axios";
+import axios from "axios/index";
 export default {
   name: "WorkerMyTask",
   data() {
