+++ src/components/worker/TaskForWorker.vue
@@ -27,7 +27,7 @@
 </template>
 
 <script>
-import axios from "axios";
+import axios from "axios/index";
 
 export default {
   name: "TaskForWorker",
